+++ home/runner/work/gradio/gradio/js/atoms/src/FullscreenButton.svelte.d.ts
@@ -1,22 +1,7 @@
-interface $$__sveltets_2_IsomorphicComponent<Props extends Record<string, any> = any, Events extends Record<string, any> = any, Slots extends Record<string, any> = any, Exports = {}, Bindings = string> {
-    new (options: import('svelte').ComponentConstructorOptions<Props>): import('svelte').SvelteComponent<Props, Events, Slots> & {
-        $$bindings?: Bindings;
-    } & Exports;
-    (internal: unknown, props: Props & {
-        $$events?: Events;
-        $$slots?: Slots;
-    }): Exports & {
-        $set?: any;
-        $on?: any;
-    };
-    z_$$bindings?: Bindings;
-}
-declare const FullscreenButton: $$__sveltets_2_IsomorphicComponent<{
-    fullscreen: any;
-}, {
-    fullscreen: CustomEvent<boolean>;
-} & {
-    [evt: string]: CustomEvent<any>;
-}, {}, {}, string>;
-type FullscreenButton = InstanceType<typeof FullscreenButton>;
+type $$ComponentProps = {
+    fullscreen: boolean;
+    onclick: (fullscreen: boolean) => void;
+};
+declare const FullscreenButton: import("svelte").Component<$$ComponentProps, {}, "">;
+type FullscreenButton = ReturnType<typeof FullscreenButton>;
 export default FullscreenButton;