+++ home/runner/work/gradio/gradio/js/atoms/src/index.d.ts
@@ -11,4 +11,6 @@ export { default as Toolbar } from "./Toolbar.svelte";
 export { default as SelectSource } from "./SelectSource.svelte";
 export { default as IconButtonWrapper } from "./IconButtonWrapper.svelte";
 export { default as FullscreenButton } from "./FullscreenButton.svelte";
+export { default as CustomButton } from "./CustomButton.svelte";
+export { default as ScrollFade } from "./ScrollFade.svelte";
 export declare const BLOCK_KEY: {};