+++ js/simpledropdown/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/simpledropdown
 
+## 0.3.35
+
+### Dependency updates
+
+- @gradio/statustracker@0.12.5
+- @gradio/utils@0.12.0
+- @gradio/atoms@0.22.2
+
 ## 0.3.34
 
 ### Dependency updates